+++ src/main/java/lotto/domain/Lottos.java
@@ -1,6 +1,5 @@
-package lotto;
+package lotto.domain;
 
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
@@ -15,12 +14,16 @@ public class Lottos {
 	public HashMap<LottoRank, Integer> getWinningStatistics(List<Integer> winningNumber) {
 		HashMap<LottoRank, Integer> result = new HashMap<>();
 
-		for (Lotto lotto : this.lottos) {
+		for (Lotto lotto : lottos) {
 			int count = lotto.getLottoNumber().getCountOfMatchLottoNumber(winningNumber);
 			result.merge(LottoRank.findLottoRankByLottoCount(count), 1, Integer::sum);
 		}
 
 		return result;
+	}
+
+	public List<Lotto> getLottos() {
+		return lottos;
 	}
 
 }
